+++ netpyne/batch/batch.py
@@ -39,7 +39,7 @@ from netpyne import specs
 
 from .utils import bashTemplate
 from .utils import createFolder
-from .grid import gridSearch
+from .grid import gridSearch, getParamCombinations, generateParamCombinations
 from .evol import evolOptim
 from .asd_parallel import asdOptim
 
@@ -97,10 +97,11 @@ class Batch(object):
     """
 
 
-    def __init__(self, cfgFile='cfg.py', netParamsFile='netParams.py', cfg=None, params=None, groupedParams=None, initCfg={}, seed=None):
+    def __init__(self, cfgFile='cfg.py', netParamsFile='netParams.py', cfg=None, netParams=None, params=None, groupedParams=None, initCfg={}, seed=None):
         self.batchLabel = 'batch_'+str(datetime.date.today())
         self.cfgFile = cfgFile
         self.cfg = cfg
+        self.netParams = netParams
         self.initCfg = initCfg
         self.netParamsFile = netParamsFile
         self.saveFolder = '/'+self.batchLabel
@@ -139,7 +140,7 @@ class Batch(object):
         createFolder(folder)
 
         # make copy of batch object to save it; but skip cfg (since instance of SimConfig and can't be copied)
-        odict = deepcopy({k:v for k,v in self.__dict__.items() if k != 'cfg'})  
+        odict = deepcopy({k:v for k,v in self.__dict__.items() if k != 'cfg' and k != 'netParams'})  
 
         if 'evolCfg' in odict:
             odict['evolCfg']['fitnessFunc'] = 'removed'
@@ -187,8 +188,15 @@ class Batch(object):
         # copy this batch script to folder, netParams and simConfig
         #os.system('cp ' + self.netParamsFile + ' ' + self.saveFolder + '/netParams.py')
 
-        netParamsSavePath = self.saveFolder+'/'+self.batchLabel+'_netParams.py'
-        os.system('cp ' + self.netParamsFile + ' ' + netParamsSavePath)
+        # if user provided a netParams object as input argument
+        if self.netParams:
+            self.netParamsSavePath = self.saveFolder+'/'+self.batchLabel+'_netParams.json'
+            self.netParams.save(self.netParamsSavePath)
+
+        # if not, use netParamsFile           
+        else:
+            self.netParamsSavePath = self.saveFolder+'/'+self.batchLabel+'_netParams.py'
+            os.system('cp ' + self.netParamsFile + ' ' + self.netParamsSavePath)
 
         os.system('cp ' + os.path.realpath(__file__) + ' ' + self.saveFolder + '/batchScript.py')
 
@@ -217,6 +225,7 @@ class Batch(object):
         self.cfg.checkErrors = False  # avoid error checking during batch
 
 
+
     def openFiles2SaveStats(self):
         stat_file_name = '%s/%s_stats.csv' %(self.saveFolder, self.batchLabel)
         ind_file_name = '%s/%s_stats_indiv.csv' %(self.saveFolder, self.batchLabel)
@@ -226,6 +235,9 @@ class Batch(object):
         individual.write('#gen  #ind  fitness  [candidate]\n')
         return stats, individual
 
+    def getParamCombinations(self):
+        if self.method in 'grid':
+            return getParamCombinations(self)
 
     def run(self):