+++ netpyne/sim/__init__.py
@@ -28,7 +28,7 @@ from neuron import h
 from .setup import initialize, setNet, setNetParams, setSimCfg, createParallelContext, readCmdLineArgs, setupRecording, setupRecordLFP, setGlobals
 
 # import run functions
-from .run import preRun, runSim, runSimWithIntervalFunc, loadBalance, calculateLFP
+from .run import preRun, runSim, runSimWithIntervalFunc, loadBalance, calculateLFP, calculateDipole
 
 # import gather functions
 from .gather import gatherData, _gatherAllCellTags, _gatherAllCellConnPreGids, _gatherCells, gatherDataFromFiles
@@ -61,6 +61,9 @@ from ..network import Network, Pop
 
 # import analysis-related module
 from .. import analysis
+
+# import plotting-related module
+from .. import plotting
 
 # import testing related functions
 from .. import tests