+++ CHANGES.md
@@ -4,7 +4,26 @@
 
 - Enabled selecting a subset of cells to save individual LFP signal from via cfg.saveLFPCells = [...] 
 
+- Added cfg.saveLFPPops to store LFP generated individually by each population
+
 **Bug fixes**
+
+- Fixed setting cfgMapping nested params when not string
+
+- Fixed defineCellShapes when sec['hObj'] is not a h.Section()
+
+- Fixed bug when using lognormal in string functions
+
+
+# Version 1.0.0.2
+
+**New features**
+
+- Added function to get list of batch grid search parameter combinations 
+
+- Save output of mpi_bulletin batches to .run and .err files
+
+- Fixed bug in loading exotic stim param values
 
 - Fixed bug in loading where sim was required to have rank attribute
 
@@ -1485,3 +1504,5 @@ First version that was uploaded to pypi. Includes following features:
 	- Parameters/specifications
 	- Instantiated networks
 	- Simulation results
+
+- Distance-based redistribution of synapses (subConn)